+++ packages/checkoutcom-integration/src/index.ts
@@ -1,0 +1,8 @@
+export { default as CheckoutcomCustomPaymentMethod } from './CheckoutcomCustomPaymentMethod';
+export {
+    default as getCheckoutcomValidationSchemas,
+    checkoutcomPaymentMethods,
+    documentPaymentMethods,
+    checkoutcomCustomPaymentMethods,
+} from './checkoutcomFieldsets/getCheckoutcomFieldsetValidationSchemas';
+export * from './checkoutcomFieldsets/CheckoutcomFormValues';