+++ packages/core/src/app/payment/getProviderWithCustomCheckout.test.ts
@@ -14,8 +14,8 @@ describe('getProviderWithCustomCheckout', () => {
     });
 
     it('returns mapped method id', () => {
-        const providerWithCustomCheckout = getProviderWithCustomCheckout(PaymentMethodId.PaypalCommerce);
-
-        expect(providerWithCustomCheckout).toEqual(PaymentMethodId.PayPalCommerceAcceleratedCheckout);
+        expect(getProviderWithCustomCheckout(PaymentMethodId.PaypalCommerce)).toEqual(PaymentMethodId.PayPalCommerceAcceleratedCheckout);
+        expect(getProviderWithCustomCheckout(PaymentMethodId.PaypalCommerceCreditCards)).toEqual(PaymentMethodId.PayPalCommerceAcceleratedCheckout);
+        expect(getProviderWithCustomCheckout(PaymentMethodId.Braintree)).toEqual(PaymentMethodId.BraintreeAcceleratedCheckout);
     });
 });