+++ packages/core/src/app/payment/getProviderWithCustomCheckout.ts
@@ -5,8 +5,12 @@ export default function getProviderWithCustomCheckout(methodId?: string | null):
         return undefined;
     }
 
-    if (methodId === PaymentMethodId.PaypalCommerce) {
+    if (methodId === PaymentMethodId.PaypalCommerce || methodId === PaymentMethodId.PaypalCommerceCreditCards) {
         return PaymentMethodId.PayPalCommerceAcceleratedCheckout;
+    }
+
+    if (methodId === PaymentMethodId.Braintree) {
+        return PaymentMethodId.BraintreeAcceleratedCheckout;
     }
 
     return methodId;