+++ packages/core/src/app/shipping/ShippingAddress.tsx
@@ -8,13 +8,14 @@ import {
     ShippingInitializeOptions,
     ShippingRequestOptions,
 } from '@bigcommerce/checkout-sdk';
-import { memoizeOne } from '@bigcommerce/memoize';
-import React, { FunctionComponent, memo, useCallback, useContext } from 'react';
+import React, { FunctionComponent, memo, useContext } from 'react';
 
+import { isPayPalFastlaneMethod, usePayPalFastlaneAddress } from '@bigcommerce/checkout/paypal-fastlane-integration';
 import { FormContext } from '@bigcommerce/checkout/ui';
 
 import ShippingAddressForm from './ShippingAddressForm';
-import StaticAddressEditable from './StaticAddressEditable';
+import { AmazonPayShippingAddress } from './AmazonPayShippingAddress';
+import { PayPalFastlaneShippingAddress } from './PayPalFastlaneShippingAddress';
 
 export interface ShippingAddressProps {
     addresses: CustomerAddress[];
@@ -49,29 +50,16 @@ const ShippingAddress: FunctionComponent<ShippingAddressProps> = (props) => {
         onAddressSelect,
         onFieldChange,
         onUseNewAddress,
-        initialize,
-        deinitialize,
         isLoading,
         shippingAddress,
         hasRequestedShippingOptions,
         addresses,
         shouldShowSaveAddress,
-        isShippingStepPending,
         isFloatingLabelEnabled,
     } = props;
 
+    const { shouldShowPayPalFastlaneShippingForm } = usePayPalFastlaneAddress();
     const { setSubmitted } = useContext(FormContext);
-
-    const initializeShipping = useCallback(
-        memoizeOne(
-            (defaultOptions: ShippingInitializeOptions) => (options?: ShippingInitializeOptions) =>
-                initialize({
-                    ...defaultOptions,
-                    ...options,
-                }),
-        ),
-        [],
-    );
 
     const handleFieldChange: (fieldName: string, value: string) => void = (fieldName, value) => {
         if (hasRequestedShippingOptions) {
@@ -81,31 +69,23 @@ const ShippingAddress: FunctionComponent<ShippingAddressProps> = (props) => {
         onFieldChange(fieldName, value);
     };
 
-    if (methodId) {
-        let options: ShippingInitializeOptions = {};
+    if (methodId === 'amazonpay' && shippingAddress) {
+        return (
+            <AmazonPayShippingAddress
+                {...props}
+                shippingAddress={shippingAddress}
+            />
+        );
+    }
 
-        if (methodId === 'amazonpay' && shippingAddress) {
-            const editAddressButtonId = 'edit-ship-button';
-
-            options = {
-                amazonpay: {
-                    editAddressButtonId,
-                },
-            };
-
-            return (
-                <StaticAddressEditable
-                    address={shippingAddress}
-                    buttonId={editAddressButtonId}
-                    deinitialize={deinitialize}
-                    formFields={formFields}
-                    initialize={initializeShipping(options)}
-                    isLoading={isShippingStepPending}
-                    methodId={methodId}
-                    onFieldChange={onFieldChange}
-                />
-            );
-        }
+    if (methodId && isPayPalFastlaneMethod(methodId) && shippingAddress && shouldShowPayPalFastlaneShippingForm) {
+        return (
+            <PayPalFastlaneShippingAddress
+                { ...props }
+                shippingAddress={shippingAddress}
+                methodId={methodId}
+            />
+        )
     }
 
     return (